server/container_manager/requirements.txt
fastapi==0.109.0
uvicorn==0.30.6
docker==7.1.0
PyYAML==6.0.1